niokr-server2/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
gunicorn==23.0.0
